requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.0
openai==1.12.0
psycopg2-binary==2.9.9
aiofiles==23.2.1
loguru==0.7.2